Update README to include Single-User Auth Info
I just upgraded to the V3 API (very nice job), but I was a little confused on authenticating. I only ever needed to authenticate as a single user, and thus didn't need to go through the hassle of building multi-user authentication. This information wasn't in this repo, so I had to go digging and found out about single-user auth on the Vimeo documentation on the API site.

I feel like this is probably a common enough occurance to warrant placing it in the readme. I retrieved all text from the Vimeo API site.